% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_fn.R
\name{generate_sub_E_fn}
\alias{generate_sub_E_fn}
\title{Pre-defined psi_star functions for sub-exponential family.}
\usage{
generate_sub_E_fn()
}
\value{
A list of pre-defined psi_star functions for sub-exponential family.
}
\description{
Pre-defined psi_star functions for sub-exponential family.
}
