% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_fn.R
\name{generate_sub_G_fn}
\alias{generate_sub_G_fn}
\title{Pre-defined psi_star functions for sub-Gaussian family.}
\usage{
generate_sub_G_fn(sig = 1)
}
\arguments{
\item{sig}{The sigma parameter of the sub-Gaussian family (default = 1).}
}
\value{
A list of pre-defined psi_star functions for sub-Gaussian family.
}
\description{
Pre-defined psi_star functions for sub-Gaussian family.
}
