% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{logSumExpTrick}
\alias{logSumExpTrick}
\title{log-sum-exp trick}
\usage{
logSumExpTrick(xs)
}
\arguments{
\item{xs}{A numeric vector.}
}
\value{
log of sum of exp of \code{xs}, which is equal to \code{log(sum(exp(xs)))}.
}
\description{
Apply log-sum-exp trick to a numeric vector.
}
