% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcp_exports.R
\name{makeGLRCUBer_}
\alias{makeGLRCUBer_}
\title{Initialize an object of GLRCUBer class}
\usage{
makeGLRCUBer_(threshold = log(1/0.05), p = 0, window_size = 100L)
}
\arguments{
\item{threshold}{stopping threshold}

\item{p}{mean of H0}

\item{window_size}{Window size for GLR computation.}
}
\value{
An object of StcpGLRCUBer class
}
\description{
Initialize an object of GLRCUBer class
}
