% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixE.R
\name{makeMixSTNormal}
\alias{makeMixSTNormal}
\title{Initialize an object of MixSTNormal class}
\usage{
makeMixSTNormal(weights = 1, log_values = 0, lambdas = 0, mu0 = 0, sig = 1)
}
\arguments{
\item{weights}{Vector of weights}

\item{log_values}{Vector of log-values}

\item{lambdas}{Vector of lambdas}

\item{mu0}{mean of H0}

\item{sig}{sigma}
}
\value{
An object of MixSTNormal class
}
\description{
Initialize an object of MixSTNormal class
}
