% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcp_exports.R
\name{makeStcpMixESRBer_}
\alias{makeStcpMixESRBer_}
\title{Initialize an object of StcpMixSRBer class}
\usage{
makeStcpMixESRBer_(threshold = log(1/0.05), weights = 1, lambdas = 1, p = 0)
}
\arguments{
\item{threshold}{stopping threshold}

\item{weights}{Vector of weights}

\item{lambdas}{Vector of lambdas}

\item{p}{mean of H0}
}
\value{
An object of StcpMixSRBer class
}
\description{
Initialize an object of StcpMixSRBer class
}
