% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcp.R
\name{Stcp}
\alias{Stcp}
\title{Stcp Class}
\description{
Stcp class supports a unified framework for sequential tests and change
detection algorithms for streams of univariate (sub-)Gaussian, binary,
and bounded random variables.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Stcp-new}{\code{Stcp$new()}}
\item \href{#method-Stcp-greet}{\code{Stcp$greet()}}
\item \href{#method-Stcp-clone}{\code{Stcp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stcp-new"></a>}}
\if{latex}{\out{\hypertarget{method-Stcp-new}{}}}
\subsection{Method \code{new()}}{
Create a new Stcp object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stcp$new(
  method = c("ST", "SR", "CU", "GLRCU"),
  family = c("Normal", "Ber", "Bounded"),
  alternative = c("two.sided", "greater", "less"),
  threshold = log(1/0.05),
  m_pre = 0,
  delta_lower = 0.1,
  delta_upper = NULL,
  k_max = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Method of the sequential procedure.
\itemize{
\item ST: Sequential test based on a mixture of E-values.
\item SR: Sequential change detection based on e-SR procedure.
\item CU: Sequential change detection based on e-CUSUM procedure.
\item GLRCU: Sequential change detection based on GLR-CUSUM procedure.
}}

\item{\code{family}}{Distribution of underlying univariate observations.
\itemize{
\item Normal: (sub-)Gaussian with sigma = 1.
\item Ber: Bernoulli distribution on \{0,1\}.
\item Bounded: General bounded distribution on [0,1]
}}

\item{\code{alternative}}{Alternative / post-change mean space
\itemize{
\item two.sided: Two-sided test / change detection
\item greater: Alternative /post-change mean is greater than null / pre-change one
\item less:  Alternative /post-change mean is less than null / pre-change one
}}

\item{\code{threshold}}{Stopping threshold. We recommend to use log(1/alpha)
for "ST" and "SR" methods where alpha is a testing level or 1/ARL.
for "CU" and "GRLCU", we recommend to tune the threshold by using
domain-specific sampler to hit the target ARL.}

\item{\code{m_pre}}{The boundary of mean parameter in null / pre-change space}

\item{\code{delta_lower}}{Minimum gap between null / pre-change space and
alternative / post-change one. It must be strictly positive for ST, SR and CU.
Currently, GLRCU does not support the minimum gap, and this param will be ignored.}

\item{\code{delta_upper}}{Maximum gap between null / pre-change space and
alternative / post-change one. It must be strictly positive for ST, SR and CU.
Currently, GLRCU does not support the maximum gap, and this param will be ignored.}

\item{\code{k_max}}{Positive integer to determine the maximum number of baselines.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Person} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stcp-greet"></a>}}
\if{latex}{\out{\hypertarget{method-Stcp-greet}{}}}
\subsection{Method \code{greet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stcp$greet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stcp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Stcp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stcp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
