% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcpR6-package.R
\docType{package}
\name{stcpR6-package}
\alias{stcpR6}
\alias{stcpR6-package}
\title{stcpR6: Sequential Test and Change-Point Detection Algorithms Based on E-Values / E-Detectors}
\description{
Algorithms of nonparametric sequential test and online change-point detection for streams of univariate (sub-)Gaussian, binary, and bounded random variables, introduced in following publications - Shin et al. (2024) \doi{10.48550/arXiv.2203.03532}, Shin et al. (2021) \doi{10.48550/arXiv.2010.08082}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shinjaehyeok/stcpR6}
  \item Report bugs at \url{https://github.com/shinjaehyeok/stcpR6/issues}
}

}
\author{
\strong{Maintainer}: Jaehyeok Shin \email{shinjaehyeok@gmail.com} (\href{https://orcid.org/0000-0003-0464-915X}{ORCID})

}
