% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcp_exports.R
\name{makeStcpMixESTNormal_}
\alias{makeStcpMixESTNormal_}
\title{Initialize an object of StcpMixSTNormal class}
\usage{
makeStcpMixESTNormal_(
  weights = 1,
  lambdas = 1,
  mu0 = 0,
  sig = 1,
  threshold = log(1/0.05)
)
}
\arguments{
\item{weights}{Vector of weights}

\item{lambdas}{Vector of lambdas}

\item{mu0}{mean of H0}

\item{sig}{sigma}

\item{threshold}{stopping threshold}
}
\value{
An object of StcpMixSTNormal class
}
\description{
Initialize an object of StcpMixSTNormal class
}
