% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcp_exports.R
\name{Stcp}
\alias{Stcp}
\title{R6 Class representing a person}
\description{
R6 Class representing a person

R6 Class representing a person
}
\details{
A person has a name and a hair color.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{First or full name of the person.}

\item{\code{hair}}{Hair color of the person.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Stcp-new}{\code{Stcp$new()}}
\item \href{#method-Stcp-set_hair}{\code{Stcp$set_hair()}}
\item \href{#method-Stcp-greet}{\code{Stcp$greet()}}
\item \href{#method-Stcp-clone}{\code{Stcp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stcp-new"></a>}}
\if{latex}{\out{\hypertarget{method-Stcp-new}{}}}
\subsection{Method \code{new()}}{
Create a new person object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stcp$new(name = NA, hair = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name.}

\item{\code{hair}}{Hair color.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Person} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stcp-set_hair"></a>}}
\if{latex}{\out{\hypertarget{method-Stcp-set_hair}{}}}
\subsection{Method \code{set_hair()}}{
Change hair color.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stcp$set_hair(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New hair color.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stcp-greet"></a>}}
\if{latex}{\out{\hypertarget{method-Stcp-greet}{}}}
\subsection{Method \code{greet()}}{
Say hi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stcp$greet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stcp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Stcp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stcp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
