% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcpR6-package.R
\docType{package}
\name{stcpR6-package}
\alias{stcpR6}
\alias{stcpR6-package}
\title{stcpR6: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
