% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcpR6-package.R
\docType{package}
\name{stcpR6-package}
\alias{stcpR6}
\alias{stcpR6-package}
\title{stcpR6: Sequential Test and Change-Point Detection Algorithms Based on E-values / E-detectors}
\description{
Algorithms of nonparametric sequential test and online change-point detection for streams of univariate (sub-)Gaussian, binary, and bounded random variables, introduced in following publications. * J. Shin, A. Ramdas, A. Rinaldo (2024) \href{https://arxiv.org/abs/2203.03532}{arXiv:2203.03532}. * J. Shin, A. Ramdas, A. Rinaldo (2021) \href{https://arxiv.org/abs/2010.08082}{arXiv:2010.08082}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shinjaehyeok/stcpR6}
  \item Report bugs at \url{https://github.com/shinjaehyeok/stcpR6/issues}
}

}
\author{
\strong{Maintainer}: Jaehyeok Shin \email{shinjaehyeok@gmail.com} (\href{https://orcid.org/0000-0003-0464-915X}{ORCID})

}
